fix AndroidLogger

+++ src/SDLApplication/AndroidLogger.cpp
@@ -15,7 +15,7 @@ namespace Mengine
     {
     }
     //////////////////////////////////////////////////////////////////////////
-    void AndroidLogger::log( ELoggerLevel _level, uint32_t _flag, const Char * _data, size_t _count )
+    void AndroidLogger::log( ELoggerLevel _level, uint32_t _flag, const Char * _data, uint32_t _count )
     {
 		MENGINE_UNUSED(_level);
         MENGINE_UNUSED(_flag);
